Fix GameSpy2 protocol query
The original GameQ PHP library [1] defines two queries for the GameSpy2
protocol, "status" respectively "players". The library also includes a
definition for Neverwinter Nights, however, that definition specifies the
GameSpy1 protocol, and NWN does not use that. Besides, the library
includes another 169 games, so adding that is a bit much.

In the meantime, GameQ has been forked [2]. The fork is hosted on
Packagist.org, but the current version, v2, appears to have removed the
NWN definition. It still includes a GameSpy2 protocol [3], and this
appears to be the same as the one from the original GameQ library although
in pure HEX format and with the "status" query renamed to "details".

This switches the query to the single "details" query defined by
GameQ 2.0.1 [3]. The "players" query seems to produce only trash
output and so is not used.

[1] http://gameq.sourceforge.net/
[2] https://github.com/Austinb/GameQ
[3] https://github.com/Austinb/GameQ/blob/2.0.1/gameq/protocols/gamespy2.php#L37